all enums converted to typesafe/enum classes, also builds in VS2012 now

+++ trunk/src/Graphics/VertexDeclaration.cpp
@@ -13,7 +13,7 @@ namespace Graphics
 		{
 			hashBuffer[i * 4 + 0] = (m_elements[i].Offset & 0xff);
 			hashBuffer[i * 4 + 1] = (m_elements[i].Offset & 0xff00) >> 8;
-			hashBuffer[i * 4 + 2] = (m_elements[i].ElementFormat & 0x7f) | ((m_elements[i].ElementUsage & 0x7f) << 4);
+			hashBuffer[i * 4 + 2] = ((int)m_elements[i].ElementFormat & 0x7f) | (((int)m_elements[i].ElementUsage & 0x7f) << 4);
 			hashBuffer[i * 4 + 3] = (m_elements[i].UsageIndex & 0xff);
 		}
